Allow the minimum 'view in admin' role of 'list_roles' to be filtered
Just because a user can list_roles does not mean they need to have the Members menu item added to their admin. For example, if we want to allow a manager to promote an employee to become a manager they need the ability to list_roles but do not need to have the Members menu added to their admin interface.